Do not warn when close fails because the file handle is bad.
Case CM-1212: The presence of this warning which was put there by
https://rt.perl.org/Public/Bug/Display.html?id=57512 means that when we do a
fastclose of file handles after fork, perl emits warnings because the file
handles were not closed/flushed properly. This is intentional on cPanel's part,
so we're turning this back off in the smallest way possible.

p5h2017 note: cPanel should keep this patch