Carp: optimize format_arg when arguments contain many references
RT #132274

This is a very minimal patch after RT discussion.
When using the CPAN version of UNIVERSAL::isa we cannot
use UNIVERSAL::isa on Carp without taking the risk of
going into one infinite loop.

As UNIVERSAL::isa on CPAN is the only one to advertise
a VERSION, we can use this value to disable the UNIVERSAL
check.

Note version bump is not required as it already occurred
since v5.27.5 release.

Signed-off-by: Nicolas R <atoomic@cpan.org>